Problem when generating code for mapping a prim. array field
MappingBuilder was generating "out.setter" as destination parameter of
statement "System.arrayCopy..." when mapping array of primitive typed
field. It is likely that this solution isn't the most proper one
(acessing vars.outFieldGetter directly from MappingBuilder), because I
found the assignments in SourceNodeVars constructor a bit confusing. In
line 50, for instance, there is an assignment of an outSetterPathFor to
the field "outField", which sounds weird. Anyway, if you provide me more
information regarding these assignments, I can provide a more elegant
solution to the problem.